src/test/java/guru/springframework/recipe/converters/Notes2NotesCommandTest.java
package guru.springframework.recipe.converters;

import guru.springframework.recipe.commands.NotesCommand;
import guru.springframework.recipe.domain.Notes;
import org.junit.Before;
import org.junit.Test;

import static org.junit.Assert.*;

public class Notes2NotesCommandTest {

    public static final Long ID_VALUE = new Long(1L);
    public static final String RECIPE_NOTES = "Notes";

    Notes2NotesCommand converter;

    @Before
    public void setUp() throws Exception {
        converter = new Notes2NotesCommand();
    }

    @Test
    public void convert() throws Exception {

        //given
        Notes notes = Notes.builder().id(ID_VALUE).recipeNotes(RECIPE_NOTES).build();

        //when
        NotesCommand notesCommand = converter.convert(notes);

        //then
        assertEquals(ID_VALUE, notesCommand.getId());
        assertEquals(RECIPE_NOTES, notesCommand.getRecipeNotes());
    }

    @Test
    public void testNull() throws Exception {
        assertNull(converter.convert(null));
    }

    @Test
    public void testEmptyObject() throws Exception {
        assertNotNull(converter.convert(new Notes()));
    }
}